fix: comments and updated with design
