+++ crowdin/messages.json
@@ -3290,6 +3290,9 @@
   "-519772806": "Smart and friendly support, 7 days a week",
   "-2015932855": "smart and friendly support",
   "-1690667791": "Why trade ETFs on Deriv",
+  "-2039405840": "The payout growth depends solely on the underlying index’s price moving tick by tick while remaining within a predefined range. The range width is determined by the growth rate and the symbol selected.",
+  "-390687537": "No, accumulator options are currently unavailable on Deriv Bot.",
+  "-2143632826": "Browse our FAQs",
   "-232157320": "Fast-paced action with durations up to 230 ticks.",
   "-844197189": "Controlled Risk",
   "-467264818": "Hone your strategies risk-free with a demo account.",
@@ -3297,17 +3300,14 @@
   "-652680277": "Step-by-step guide to trade on accumulator options",
   "-712487889": "Read: Guide to trading accumulator options",
   "-1923211546": "Open your first accumulator options trade with our step-by-step guide.",
+  "-1833614673": "How to trade accumulator options",
   "-1627820643": "Move fast,<br/>when the market doesn’t",
-  "-360691661": "Looking for a fast-paced, strategic and potentially explosive way to trade?",
   "-244980648": "What are accumulator options?",
   "-1356285247": "Accumulator options icon",
   "-1500514644": "Accumulator",
   "-1167572902": "Secure your gains with manual or automatic profit-taking. With durations of up to 230 ticks, accumulator options are ideal for those looking for fast-paced, strategic trades while balancing risk and reward.",
   "-677391007": "Deriv Go",
-  "-1833614673": "How to trade accumulator options",
-  "-2039405840": "The payout growth depends solely on the underlying index’s price moving tick by tick while remaining within a predefined range. The range width is determined by the growth rate and the symbol selected.",
-  "-390687537": "No, accumulator options are currently unavailable on Deriv Bot.",
-  "-2143632826": "Browse our FAQs",
+  "-360691661": "Looking for a fast-paced, strategic and potentially explosive way to trade?",
   "-1759381505": "What are turbo options?",
   "-1461784880": "How do vanilla options work?",
   "-231290085": "Start connecting with fellow traders to make deposits and withdrawals.",
